// src/app/shared/services/model/model.ts
export interface pageSelection {
  skip: number;
  limit: number;
}
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
  export interface userWallet {
    id: number;
    refId: string;
    transactionFor: string;
    date: string;
    total: string;
    status: string;
    isSelected: boolean;
  }